+++ src/charts/Legend.jsx
@@ -1,21 +1,21 @@
 import React from "react";
-import {ShallowComponent, Generator, Class, Arrays, Maps} from "robe-react-commons";
+import ShallowComponent from "robe-react-commons/lib/components/ShallowComponent";
 import "./Legend.css";
-import ThumbnailGroup from "../layouts/ThumbnailGroup";
-import ThumbnailItem from "../layouts/ThumbnailItem";
 
 export default class Legend extends ShallowComponent {
 
-    static propTypes: Map = {
+    static propTypes = {
         className: React.PropTypes.string,
         width: React.PropTypes.number,
+        data: React.PropTypes.array,
     };
 
     static defaultProps = {
-        width: 500
+        width: 500,
+        data: []
     };
 
-    constructor(props: Object) {
+    constructor(props) {
         super(props)
     }
 
@@ -24,27 +24,22 @@ export default class Legend extends ShallowComponent {
             <div
                 className={this.props.className}
                 style={{width: this.props.width, marginTop: 20}}>
-                <ThumbnailGroup>
+                <div className="rb-thumbnail">
                     {this.__renderLegend()}
-                </ThumbnailGroup>
+                </div>
             </div>
         )
     }
 
-    __renderLegend() {
+    __renderLegend = () => {
         let legends = this.props.data, arr = [];
         for (let key in legends) {
             if (legends.hasOwnProperty(key)) {
                 let lengend = legends[key];
-
                 arr.push(
-                    <ThumbnailItem
-                        key={key}
-                        className="legend-item">
-                        <div style={{borderColor: lengend.fill, color: lengend.fill}}>
-                            {lengend.label}
-                        </div>
-                    </ThumbnailItem>)
+                    <div style={{borderColor: lengend.fill, color: lengend.fill}}>
+                        {lengend.label}
+                    </div>)
             }
         }
         return arr;