+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## 1.1.27 -- Developing...
+* Bug fixed the Legend for Chart's components
 
 ## 1.1.26
 * New prop "onAfterChange" added to Slider component.